toutiao-backend/requirements.txt
aliyun-python-sdk-core-v3==2.11.0
amqp==2.3.2
aniso8601==4.0.1
APScheduler==3.5.3
billiard==3.5.0.4
blinker==1.4
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dnspython==1.16.0
elasticsearch==6.3.1
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Limiter==1.0.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
greenlet==0.4.15
grpcio==1.16.1
grpcio-tools==1.16.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kafka-python==1.4.5
kombu==4.2.1
limits==1.3
MarkupSafe==1.1.0
monotonic==1.5
mysqlclient==1.3.13
protobuf==3.6.1
pycryptodome==3.7.0
PyJWT==1.6.4
pymongo==3.7.2
python-engineio==3.1.2
python-socketio==3.0.0
pytz==2018.7
qiniu==7.2.2
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.21.0
six==1.11.0
SQLAlchemy==1.2.14
tzlocal==1.5.1
urllib3==1.24.1
vine==1.1.4
Werkzeug==0.14.1
